+++ packages/load-config/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sveltejs/load-config
 
+## 0.2.0
+
+### Minor Changes
+
+-   feat: make it possible to pass a config file path directly ([#3066](https://github.com/sveltejs/language-tools/pull/3066))
+
+### Patch Changes
+
+-   fix: load esm version of Vite ([#3065](https://github.com/sveltejs/language-tools/pull/3065))
+
 ## 0.1.1
 
 ### Patch Changes